move ThemeContext.js to styles folder
