+++ client/src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 
 import "./index.css";
 import App from "./components/App";
-import { ThemeProvider } from "./ThemeContext";
+import { ThemeProvider } from "./styles/ThemeContext";
 import { createRoot } from "react-dom/client";
 
 import { AuthProvider } from "@propelauth/react";
